Add getModel() to LwM2mObjectTree.
